Culminacion de registro, perfil, panel de contro. Obtencion de datos desde Servidor en la nube
