--- src/modules/AdminPanel/CubicleManagement/CubicleManagement.css
@@ -1,42 +1,0 @@
-.cubicle-management {
-    background: white;
-    padding: 2rem;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .cubicle-management h2 {
-    margin-bottom: 1.5rem;
-    font-size: 1.5rem;
-    color: #333;
-  }
-  
-  .cubicle-management table {
-    width: 100%;
-    border-collapse: collapse;
-  }
-  
-  .cubicle-management th, .cubicle-management td {
-    padding: 0.75rem;
-    text-align: left;
-    border: 1px solid #ddd;
-  }
-  
-  .cubicle-management th {
-    background-color: #f8f9fa;
-  }
-  
-  .cubicle-management button {
-    margin-right: 0.5rem;
-    padding: 0.5rem 1rem;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .cubicle-management button:hover {
-    background-color: #0056b3;
-  }
-
