--- src/modules/AdminPanel/Dashboard/Dashboard.jsx
@@ -1,52 +1,0 @@
-import React from 'react';
-import { Bar, Line } from 'react-chartjs-2';
-import { Chart as ChartJS, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend, PointElement, LineElement } from 'chart.js';
-import './Dashboard.css';
-
-ChartJS.register(CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend, PointElement, LineElement);
-
-const Dashboard = () => {
-  // Datos de ejemplo para el gráfico de barras
-  const barData = {
-    labels: ['Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio'],
-    datasets: [
-      {
-        label: 'Reservas',
-        data: [12, 19, 3, 5, 2, 3],
-        backgroundColor: 'rgba(75, 192, 192, 0.2)',
-        borderColor: 'rgba(75, 192, 192, 1)',
-        borderWidth: 1,
-      },
-    ],
-  };
-
-  // Datos de ejemplo para el gráfico de líneas
-  const lineData = {
-    labels: ['Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio'],
-    datasets: [
-      {
-        label: 'Usuarios nuevos',
-        data: [10, 15, 9, 14, 20, 25],
-        fill: false,
-        backgroundColor: 'rgba(153, 102, 255, 0.2)',
-        borderColor: 'rgba(153, 102, 255, 1)',
-      },
-    ],
-  };
-
-  return (
-    <div className="dashboard">
-      <h2>Dashboard</h2>
-      <div className="chart-container">
-        <h3>Reservas por Mes</h3>
-        <Bar data={barData} />
-      </div>
-      <div className="chart-container">
-        <h3>Usuarios Nuevos por Mes</h3>
-        <Line data={lineData} />
-      </div>
-    </div>
-  );
-};
-
-export default Dashboard;
