--- src/modules/AdminPanel/NotificationHistory/NotificationHistory.css
@@ -1,28 +1,0 @@
-.notification-history {
-    background: white;
-    padding: 2rem;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .notification-history h2 {
-    margin-bottom: 1.5rem;
-    font-size: 1.5rem;
-    color: #333;
-  }
-  
-  .notification-history table {
-    width: 100%;
-    border-collapse: collapse;
-  }
-  
-  .notification-history th, .notification-history td {
-    padding: 0.75rem;
-    text-align: left;
-    border: 1px solid #ddd;
-  }
-  
-  .notification-history th {
-    background-color: #f8f9fa;
-  }
-
