--- src/modules/AdminPanel/UserManagement/UserManagement.css
@@ -1,41 +1,0 @@
-.user-management {
-  background: white;
-  padding: 2rem;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.user-management h2 {
-  margin-bottom: 1.5rem;
-  font-size: 1.5rem;
-  color: #333;
-}
-
-.user-management table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.user-management th, .user-management td {
-  padding: 0.75rem;
-  text-align: left;
-  border: 1px solid #ddd;
-}
-
-.user-management th {
-  background-color: #f8f9fa;
-}
-
-.user-management button {
-  margin-right: 0.5rem;
-  padding: 0.5rem 1rem;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.user-management button:hover {
-  background-color: #0056b3;
-}
