+++ src/modules/Header/Header.jsx
@@ -1,28 +1,22 @@
-import React, { useState } from 'react';
+import React, { useState, useContext } from 'react';
+import { useNavigate } from 'react-router-dom';
 import MenuIcon from '@mui/icons-material/Menu';
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import './Header.css';
+import AuthContext from '../../context/AuthContext';
 
 const Header = () => {
-
-
-  const user = {
-    avatar: 'https://via.placeholder.com/100', // Reemplaza con la URL de la imagen del avatar del usuario
-    name: 'John Doe',
-    email: 'johndoe@example.com',
-    bio: 'Software engineer with a passion for developing innovative programs that expedite the efficiency and effectiveness of organizational success.'
-  };
-
-  const handleLogout = () => {
-    // Aquí puedes manejar el cierre de sesión del usuario
-    console.log('User logged out');
-  };
-
-
+  const { user, logout } = useContext(AuthContext);
+  const navigate = useNavigate();
   const [menuOpen, setMenuOpen] = useState(false);
 
   const toggleMenu = () => {
     setMenuOpen(!menuOpen);
+  };
+
+  const handleLogout = () => {
+    logout();
+    navigate('/login');
   };
 
   return (
@@ -36,14 +30,14 @@ const Header = () => {
             <ul>
               <li><a href="/">Home</a></li>
               <li><a href="/profile">Profile</a></li>
-              <li><a href="/settings">Settings</a></li>
+              <li><a href="/user">Panel</a></li>
             </ul>
           </div>
         )}
       </div>
       <div className="header-right">
         <NotificationsIcon className="header-icon" />
-        <img src={user.avatar} alt="User Avatar" className="header-avatar" />
+        <img src={user?.avatar || 'https://via.placeholder.com/100'} alt="User Avatar" className="header-avatar" />
         <button className="header-logout-button" onClick={handleLogout}>Logout</button>
       </div>
     </header>
