+++ src/modules/Login/Login.css
@@ -2,21 +2,65 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 100%;
-  background-image: url("./../../assets/images/fondo_main1.jpg");
+  height: 100vh;
+  overflow: hidden;
+  position: relative;
   background-position: center;
   background-repeat: no-repeat;
   background-attachment: fixed;
   background-size: cover;
 }
 
+
+
+/* ESTILOS DEL CONTENEDOR DEL VIDEO DE FONDO */
+.login-video {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 0;
+  background: #000000;
+}
+
+.login-video video{
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  z-index: 0;
+}
+
+.login-video .bgdeco {
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  position: absolute;
+  background: linear-gradient(to right, #000 -30%, rgba(0, 0, 0, 0) 70%);
+  z-index: 1;
+}
+
+.login-video .bgdeco.q{
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  position: absolute;
+  background: linear-gradient(to left, #000 -30%, rgba(0, 0, 0, 0) 70%);
+  z-index: 1;
+}
+
+
+
+/* ESTILOS DEL CONTENEDOR DEL LOGO DE LA EMPRESA */
 .login-logo {
   width: auto;
-  height: 10rem;
+  height: 6rem;
   display: flex;
   position: absolute;
-  top: 0.75rem;
-  left: 0.75rem;
+  top: 2rem;
+  left: 2rem;
 }
 
 .login-logo img {
@@ -25,18 +69,36 @@
   object-fit: cover;
 }
 
+
+
+/* ESTILOS DEL CONTENEDOR DEL FORMULARIO */
 .login-form {
   width: 100%;
   max-width: 24rem;
   height: 100%;
-  max-height: 28rem;
+  max-height: 30.5rem;
   position: relative;
+  overflow: hidden;
   padding: 2.5rem;
   border-radius: 10px;
   border: 1px solid #ddd;
   text-align: center;
-  background: rgba(255, 255, 255, 1);
+  background: rgba(255, 255, 255, 0.6);
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.5);
+}
+
+
+
+/* ESTILOS DEL CONTENIDO Y ELEMENTOS DEL FORMULARIO */
+.login-form .form-content {
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  overflow: hidden;
+  position: absolute;
+  padding: 2.5rem;
+  z-index: 2;
 }
 
 .login-form h2 {
@@ -73,6 +135,9 @@
   background-color: #1d1f20;
 }
 
+
+
+/* ESTILOS DEL CONTENEDOR DEL LINK PARA CAMBIO A SIGN-UP */
 .login-form .change-form {
   width: 100%;
   height: 15%;
@@ -80,10 +145,10 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  background: #ffffff;
   left: 0;
   bottom: 0;
 }
-
 
 .login-form p {
   font-size: 0.9rem;
@@ -98,3 +163,27 @@
 .login-form a:hover {
   text-decoration: underline;
 }
+
+
+
+/* ESTILOS DEL EFECTO BLUR DEL FONDO */
+.login-form .blur-effect {
+  content: "";
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  overflow: hidden;
+  position: absolute;
+  padding: 2.5rem;
+  border-radius: 10px;
+  border: 1px solid #ddd;
+  text-align: center;
+  background: rgba(255, 255, 255, 0.5);
+  backdrop-filter: blur(10px);
+  filter: blur(200px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.5);
+  z-index: 1;
+}
+
+
