+++ src/modules/Panel/Panel.css
@@ -1,0 +1,137 @@
+.panel {
+  display: flex;
+  height: 100vh;
+  background: #e9ecef;
+}
+
+.panel .panel-nav {
+  width: 16rem;
+  background: white;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.panel .panel-nav .panel-logo {
+  width: calc(100% - 2rem);
+  height: auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  flex-direction: column;
+  margin-bottom: 1rem;
+  padding-bottom: 1rem;
+  margin: 1rem;
+  border-bottom: 1px solid #ddd;
+}
+
+.panel .panel-nav .panel-logo img {
+  width: 50%;
+  max-width: 50%;
+  height: auto;
+  object-fit: contain;
+  aspect-ratio: 4/3;
+}
+
+.panel .panel-nav .panel-logo p {
+  width: 50%;
+  max-width: 50%;
+  font-weight: bold;
+}
+
+.panel .panel-nav .panel-options {
+  width: 100%;
+  display: flex;
+  flex-grow: 1;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: space-between;
+  overflow: hidden;
+  overflow-y: auto;
+}
+
+.panel .panel-nav .panel-options ul {
+  width: 100%;
+  padding: 0;
+  padding: 0rem 1rem;
+  list-style: none;
+}
+
+.panel .panel-nav .panel-options li {
+  width: 100%;
+  padding: 1rem;
+  cursor: pointer;
+  border-radius: 5px;
+  margin-bottom: 0.5rem;
+  transition: background 0.3s, transform 0.15s;
+  overflow: hidden;
+  user-select: none;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.panel .panel-nav .panel-options li:hover {
+  background: #e6e6e6;
+}
+
+.panel .panel-nav .panel-options li:active {
+  transform: scale(0.9);
+}
+
+.panel .panel-nav .panel-options li.active {
+  background: #656d75;
+  color: white;
+}
+
+.panel .panel-nav .panel-user {
+  width: calc(100% - 2rem);
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-top: 1px solid #ddd;
+  margin: 0rem 1rem;
+  padding: 1rem;
+}
+
+.panel .panel-nav .panel-user img {
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  object-fit: cover;
+  aspect-ratio: 1/1;
+}
+
+.panel .panel-nav .panel-user button {
+  width: 3rem;
+  height: 3rem;
+  right: 0;
+  border-radius: 50%;
+  border: none;
+  overflow: hidden;
+  position: absolute;
+  display: flex;
+  background: none;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+}
+
+.panel .panel-nav .panel-user button > *{
+  transform: scale(0.95);
+}
+
+.panel .panel-nav .panel-user button:active > *{
+  transform: scale(0.85);
+}
+
+.panel .panel-content {
+  flex: 1;
+  padding: 2rem;
+  background: #f1f3f5;
+  overflow: hidden;
+  overflow-y: auto;
+}
