+++ src/modules/Panel/optionsAdmin/Dashboard/Dashboard.css
@@ -1,0 +1,7 @@
+.dashboard {
+    background: white;
+    padding: 2rem;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+
