+++ src/modules/Panel/optionsAdmin/NotificationForm/NotificationForm.css
@@ -1,0 +1,50 @@
+.notification-form {
+    background: white;
+    padding: 2rem;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+  
+  .notification-form h2 {
+    margin-bottom: 1.5rem;
+    font-size: 1.5rem;
+    color: #333;
+  }
+  
+  .notification-form form {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .notification-form input,
+  .notification-form textarea {
+    display: block;
+    width: calc(100% - 2rem);
+    padding: 0.75rem;
+    margin-bottom: 1.5rem;
+    border: 1px solid #ddd;
+    border-radius: 5px;
+    margin: 0 auto 1.5rem auto;
+  }
+  
+  .notification-form textarea {
+    height: 100px;
+    resize: none;
+  }
+  
+  .notification-form button {
+    display: block;
+    width: calc(100% - 2rem);
+    padding: 0.75rem;
+    background-color: #007bff;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    margin: 0 auto;
+  }
+  
+  .notification-form button:hover {
+    background-color: #0056b3;
+  }
+
