+++ src/modules/Panel/optionsAdmin/ReservationManagement/ReservationManagement.css
@@ -1,0 +1,41 @@
+.reservation-management {
+  background: white;
+  padding: 2rem;
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.reservation-management h2 {
+  margin-bottom: 1.5rem;
+  font-size: 1.5rem;
+  color: #333;
+}
+
+.reservation-management table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.reservation-management th, .reservation-management td {
+  padding: 0.75rem;
+  text-align: left;
+  border: 1px solid #ddd;
+}
+
+.reservation-management th {
+  background-color: #f8f9fa;
+}
+
+.reservation-management button {
+  margin-right: 0.5rem;
+  padding: 0.5rem 1rem;
+  background-color: #007bff;
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.reservation-management button:hover {
+  background-color: #0056b3;
+}
