--- src/modules/ReservationForm/ReservationForm.css
@@ -1,51 +1,0 @@
-.reservation-container {
-    display: flex;
-    justify-content: space-around;
-    align-items: flex-start;
-    padding-top: 60px;
-    background: #e9ecef;
-    height: 100vh;
-  }
-  
-  .reservation-form {
-    background: white;
-    padding: 2.5rem;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    text-align: center;
-    width: 100%;
-    max-width: 400px;
-  }
-  
-  .reservation-form h2 {
-    margin-bottom: 1.5rem;
-    font-size: 1.5rem;
-    color: #333;
-  }
-  
-  .reservation-form input {
-    display: block;
-    width: calc(100% - 2rem);
-    padding: 0.75rem;
-    margin-bottom: 1.5rem;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    margin: 0 auto 1.5rem auto;
-  }
-  
-  .reservation-form button {
-    display: block;
-    width: calc(100% - 2rem);
-    padding: 0.75rem;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    margin: 0 auto;
-  }
-  
-  .reservation-form button:hover {
-    background-color: #0056b3;
-  }
-
