+++ src/modules/SignUp/SignUp.jsx
@@ -16,7 +16,7 @@ const SignUp = () => {
       return;
     }
 
-    fetch('http://localhost/POST/signup.php', {
+    fetch('https://localdbs.com/POST/signup.php', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
@@ -40,38 +40,49 @@ const SignUp = () => {
 
   return (
     <div className="signup-container">
+      <div className="signup-video">
+        <video src="src\assets\videos\video.mp4" autoPlay loop></video>
+        <div className="bgdeco q"></div>
+        <div className="bgdeco w"></div>
+        <div className="bgdeco e"></div>
+        <div className="bgdeco r"></div>
+        <div className="bgdeco t"></div>
+      </div>
       <div className="signup-logo">
-        <img src="src\assets\images\UGB_LOGOTIPO_HORIZONTAL.png" alt="Logo UGB" />
+        <img src="src\assets\images\ugb_icon_hor_light.38b2c33c.png" alt="Logo UGB" />
       </div>
       <form className="signup-form" onSubmit={handleSubmit}>
-        <h2>Sign Up</h2>
-        <input
-          type="email"
-          placeholder="Email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          required
-        />
-        <input
-          type="password"
-          placeholder="Password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          required
-        />
-        <input
-          type="password"
-          placeholder="Confirm Password"
-          value={confirmPassword}
-          onChange={(e) => setConfirmPassword(e.target.value)}
-          required
-        />
-        <button type="submit">Submit</button>
-        <div className="change-form">
-          <p>
-            Already have an account? <Link to="/login">Login</Link>
-          </p>
+        <div className="form-content">
+          <h2>Registrarse</h2>
+          <input
+            type="email"
+            placeholder="Email"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            required
+          />
+          <input
+            type="password"
+            placeholder="Password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            required
+          />
+          <input
+            type="password"
+            placeholder="Confirm Password"
+            value={confirmPassword}
+            onChange={(e) => setConfirmPassword(e.target.value)}
+            required
+          />
+          <button type="submit">Submit</button>
+          <div className="change-form">
+            <p>
+            ¿Ya tienes una cuenta? <Link to="/login">Inicia sesión aquí</Link>
+            </p>
+          </div>
         </div>
+        <div className="blur-effect"></div>
       </form>
     </div>
   );
