--- src/modules/UserPanel/UserPanel.css
@@ -1,37 +1,0 @@
-.user-panel {
-    display: flex;
-    height: 100vh;
-    background: #e9ecef;
-  }
-  
-  .user-nav {
-    width: 250px;
-    background: white;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    padding: 1rem;
-  }
-  
-  .user-nav ul {
-    list-style: none;
-    padding: 0;
-  }
-  
-  .user-nav li {
-    padding: 1rem;
-    cursor: pointer;
-    border-radius: 5px;
-    margin-bottom: 0.5rem;
-    transition: background 0.3s;
-  }
-  
-  .user-nav li:hover {
-    background: #f8f9fa;
-  }
-  
-  .user-content {
-    flex: 1;
-    padding: 2rem;
-    background: #f1f3f5;
-    overflow-y: auto;
-  }
-
