--- src/modules/UserProfile/UserProfile.css
@@ -1,61 +1,0 @@
-.profile-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
-  }
-  
-  .profile-card {
-    background: white;
-    padding: 2.5rem;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    text-align: center;
-    width: 100%;
-    max-width: 400px;
-  }
-  
-  .profile-avatar {
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-    margin-bottom: 1rem;
-  }
-  
-  .profile-title {
-    margin-bottom: 1.5rem;
-    font-size: 1.5rem;
-    color: #333;
-  }
-  
-  .profile-input {
-    display: block;
-    width: calc(100% - 2rem);
-    padding: 0.75rem;
-    margin-bottom: 1.5rem;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    margin: 0 auto 1.5rem auto;
-    background-color: #f8f9fa;
-    color: #333;
-  }
-  
-  .profile-input.profile-bio {
-    resize: none;
-    height: 80px;
-  }
-  
-  .profile-edit-button {
-    display: inline-block;
-    padding: 0.75rem 1.5rem;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .profile-edit-button:hover {
-    background-color: #0056b3;
-  }
-
